Idea: roads
Roads are interesting because they change the topology of the world.

An empty plane is the same everywhere, but with roads there's a huge
difference between being in a desert, and being next to a road that can
take you anywhere.

Roads are also natural waypoints. They provide direction.

Rivers are like roads, but you also know which way is forward.
(Though in a 2D game you always know which way is 'north'.)